+++ src/main/java/domain/constant/Country.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.constant;
 
 public enum Country {
     CHO("초나라"),
