+++ src/main/java/domain/state/ChoTurn.java
@@ -1,4 +1,6 @@
-package domain;
+package domain.state;
+
+import domain.constant.Country;
 
 public class ChoTurn implements State {
     @Override
