+++ src/main/java/domain/state/State.java
@@ -1,4 +1,6 @@
-package domain;
+package domain.state;
+
+import domain.constant.Country;
 
 public interface State {
     State changeTurn();
